+++ resource/organize.cpp
@@ -79,9 +79,9 @@ string resource_organize (void * webserver_request)
   
   
   if (request->query.count ("remove")) {
-    size_t remove = convert_to_int (request->query["remove"]);
+    int remove = convert_to_int (request->query["remove"]);
     vector <string> resources = request->database_config_user()->getActiveResources ();
-    if (remove < resources.size ()) {
+    if (remove < static_cast<int>(resources.size ())) {
       resources.erase (resources.begin () + remove);
     }
     request->database_config_user()->setActiveResources (resources);
@@ -93,8 +93,8 @@ string resource_organize (void * webserver_request)
   if (!movefrom.empty ()) {
     string moveto =  request->post ["moveto"];
     if (!moveto.empty ()) {
-      int from = convert_to_int (movefrom);
-      int to = convert_to_int (moveto);
+      size_t from = static_cast<size_t> (convert_to_int (movefrom));
+      size_t to = static_cast<size_t>(convert_to_int (moveto));
       vector <string> resources = request->database_config_user()->getActiveResources ();
       array_move_from_to (resources, from, to);
       request->database_config_user()->setActiveResources (resources);
